backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
python-dotenv
psycopg2-binary
sqlalchemy
boto3
